Refactored HobbyBubble and HobbyTag components to use DynamicBackground wrapper for cleaner structure and reusability
