+++ app/components/DynamicBackground/style.module.scss
@@ -1,0 +1,15 @@
+@use '@/styles/_vars.scss' as *;
+
+@use 'sass:color';
+
+.dynamicBackground {
+    display: inline-block;
+    border: 1px solid var(--darkerBorderColor);
+    border-radius: 10px;
+    background: linear-gradient( -45deg, var(--baseColor) 10%, var(--lighterColor) 100%);
+    .Image {
+        &:hover {
+          cursor: pointer;
+        }
+      }
+}
