+++ app/components/HobbyBubble/style.module.scss
@@ -3,22 +3,20 @@
 @use 'sass:color';
 
 .hobbyBubble {
-    display: flex;
-    flex-direction: row;
-    align-items: end;
-    justify-content: space-between;
-    padding: 10px;
-    border: 1px solid var(--darkerBorderColor);
-    border-radius: 10px;
-    width: 100px;
-    aspect-ratio: 3/2;
-    gap: 10px;
-    font-size: 18px;
-    color: black;
-    background: linear-gradient( -45deg, var(--baseColor) 10%, var(--lighterColor) 100%);
-    .Image {
-        &:hover {
-          cursor: pointer;
-        }
-      }
+  display: flex;
+  flex-direction: row;
+  align-items: end;
+  justify-content: space-between;
+  padding: 10px;
+  width: 100px;
+  aspect-ratio: 3/2;
+  gap: 10px;
+  font-size: 18px;
+  color: $bubbleTextColor;
+
+  .Image {
+    &:hover {
+      cursor: pointer;
+    }
+  }
 }
