+++ app/components/ImageWithTitle/index.jsx
@@ -1,12 +1,16 @@
-import styles from "./ImageWithTitle.module.scss";
+import React from "react";
+import styles from "./style.module.scss";
 
-export default function ImageWithTitle({ imageUrl, title, subtitle }) {
+const ImageWithTitle = ({ imageUrl, title, subtitle,onClick }) => {
   return (
     <div
       className={styles.imageWithTitle}
       style={{
         backgroundImage: `url(${imageUrl})`,
       }}
+      onClick={onClick} 
+      role={onClick ? "button" : undefined} 
+      tabIndex={onClick ? 0 : undefined} 
     >
       <div className={styles.content}>
         <h1>{title}</h1>
@@ -14,4 +18,6 @@ export default function ImageWithTitle({ imageUrl, title, subtitle }) {
       </div>
     </div>
   );
-}
+};
+
+export default ImageWithTitle;
