+++ app/test/page.jsx
@@ -1,4 +1,3 @@
-"use client";
 import React from 'react'
 import styles from './style.module.scss'
