+++ styles/_vars.scss
@@ -16,6 +16,7 @@ $grayBackground: var(--gray);
 $lightblueBorder: var(--lightblue);
 $textColor: var(--darkGray);
 $buttonColor: var(--mutedBlue);
+$bubbleTextColor: var(--mutedBlue);
 $placeHolderColor: var(--lightGray);
 
 //Gradients
